+++ scripts/mcdo.sh
@@ -16,6 +16,7 @@ outs=$1 ; shift
 var=$1 ; shift
 period=$1 ; shift
 region=$1 ; shift
+alias=$1 ; shift
 
 tmp=~/tmp/$(basename $0)
 mkdir -p $tmp
@@ -31,7 +32,10 @@ if [ "$region" ] ; then
     lonmax=$(echo $region | cut -d "," -f 4)
     selregion="-sellonlatbox,$lonmin,$lonmax,$latmin,$latmax"
 fi
-
+if [ "$alias" ] ; then 
+    IFS=", " read var filevar scale offset <<< $alias 
+    selalias=-expr,"$var=${filevar}*${scale}+${offset};"
+fi
 i=0
 # Loop on the list of list of files 
 for out in $outs ; do 
@@ -41,19 +45,23 @@ for out in $outs ; do
     eval files=\$$i
     vfiles=""
 
-    # Make all selections (time, space, variable) before applying the operator
+    # Make all selections (aliasing, time, space, variable) before applying the operator
     for file in $files ; do
 	tmp2=$tmp/$(basename $file) ; rm -f $tmp2
-	if [ "$period" ] ; then 
-	    cdo ${selvar#-} $selregion $seldate $file $tmp2  && [ -f $tmp2 ] && vfiles+=" "$tmp2 
+	if [ $alias ] ; then 
+	    cdo ${selalias#-} ${selvar} $selregion $seldate $file $tmp2  && [ -f $tmp2 ] && vfiles+=" "$tmp2 
 	else
-	    if [ "$var" ] ; then 
-		cdo ${selvar#-} $selregion $file $tmp2 && [ -f $tmp2 ] && vfiles+=" "$tmp2
-	    else 
-		if [ "$region" ] ; then 
-		    cdo ${selregion#-} $file $tmp2 && [ -f $tmp2 ] && vfiles+=" "$tmp2
-		else
-		    vfiles+=" "$file ; 
+	    if [ "$period" ] ; then 
+		cdo ${selvar#-} $selregion $seldate $file $tmp2  && [ -f $tmp2 ] && vfiles+=" "$tmp2 
+	    else
+		if [ "$var" ] ; then 
+		    cdo ${selvar#-} $selregion $file $tmp2 && [ -f $tmp2 ] && vfiles+=" "$tmp2
+		else 
+		    if [ "$region" ] ; then 
+			cdo ${selregion#-} $file $tmp2 && [ -f $tmp2 ] && vfiles+=" "$tmp2
+		    else
+			vfiles+=" "$file ; 
+		    fi
 		fi
 	    fi
 	fi
@@ -62,7 +70,7 @@ for out in $outs ; do
     # Then, assemble all datafiles in a single one before applying the operator 
     # (because it may be non-linear in time - e.g. eigenvectors )
     if [ "$vfiles" ] ; then 
-	tmp3=$tmp/$(basename $0).nc ; rm $-f tmp3
+	tmp3=$tmp/$(basename $0).nc ; rm -f $tmp3
         # let us avoid single file copy followed by rm ...
 	if [ $(echo $vfiles | wc -w) -gt 1 ] ; then 
 	    cdo copy $vfiles $tmp3 && [ "$var" -o "$period" -o "$region" ] && rm $vfiles