+++ testing/test_examples.sh
@@ -30,4 +30,7 @@ scripts=$scripts" clean_cache.py"
 
 echo "tested example scripts : "$scripts 
 export CLIMAF_LOG_LEVEL=critical
+
 python $thisdir/iterate_python_scripts.py -V -f $scripts
+#for s in $scripts ; do cdat $s ; done
+